# backend/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Captapi API",
    "description": "Extract transcripts, summaries, comments and stats from social media videos (YouTube, TikTok, Instagram, Facebook). Connect by AI agents and no-code tools via the MCP server (@captapi/mcp), the CLI (@captapi/cli), the n8n community node (n8n-nodes-captapi), the Make.com custom app, or the Apify Actor (BYO key). Full guide: https://captapi.com/docs/integrations",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/truth-social/profile": {
      "get": {
        "tags": [
          "Truth Social"
        ],
        "summary": "Truth Social profile",
        "operationId": "profile_v1_truth_social_profile_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Truth Social profile URL or @username",
              "title": "Url"
            },
            "description": "Truth Social profile URL or @username"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/truth-social/user-posts": {
      "get": {
        "tags": [
          "Truth Social"
        ],
        "summary": "Truth Social user posts",
        "operationId": "user_posts_v1_truth_social_user_posts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Truth Social profile URL or @username",
              "title": "Url"
            },
            "description": "Truth Social profile URL or @username"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 80,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/truth-social/post": {
      "get": {
        "tags": [
          "Truth Social"
        ],
        "summary": "Truth Social post",
        "operationId": "post_v1_truth_social_post_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Truth Social post URL or post ID",
              "title": "Url"
            },
            "description": "Truth Social post URL or post ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Your API key (capt_live_… / capt_test_…) or a dashboard session token.",
        "scheme": "bearer"
      }
    }
  }
}